Move ImageWellView to its own file
